test: Refactor async hook installation
